Added Drive with Turn(WIP) and cleaned up WriteLines
Added setup function to ChangeArmPosition that takes an int
Replaced a DriveToPosition in DriveToWrench with DriveToPositionWithHeading
Cleaned up uncessary WriteLines in ReadButtonLight
Modified DriveWithTurnSpeed Algorithim

+++ FEHRobot/Subsystems/DriveTrain.h
@@ -15,7 +15,7 @@ private:
 public:
     DriveTrain();
 
-    int driveAndTurn(int heading, int power, int turnSpeed);
+    int driveAndTurn(int heading, int power, float turnSpeed);
     int turn(int power);
     int drive(int heading, int power);
     int stop();
